connection refatoring
