+++ src/background/rpcHandler.ts
